+++ program/rust/Cargo.toml
@@ -10,7 +10,6 @@ bindgen = "0.60.1"
 
 [dependencies]
 solana-program = "=1.10.29"
-borsh = "0.9"
 bytemuck = "1.11.0"
 thiserror = "1.0"
