+++ program/rust/build.rs
@@ -4,15 +4,8 @@ use bindgen::Builder;
 fn main() {
     println!("cargo:rustc-link-search=./program/c/target");
 
-    let borsh_derives = ["BorshSerialize".to_string(), "BorshDeserialize".to_string()];
-
     //make a parser and to it type, traits pairs
-    let mut parser = build_utils::DeriveAdderParserCallback::new();
-    parser.register_traits("cmd_hdr", borsh_derives.to_vec());
-    parser.register_traits("pc_acc", borsh_derives.to_vec());
-    parser.register_traits("pc_price_info", borsh_derives.to_vec());
-    parser.register_traits("cmd_upd_price", borsh_derives.to_vec());
-    parser.register_traits("pc_ema", borsh_derives.to_vec());
+    let parser = build_utils::DeriveAdderParserCallback::new();
 
     //generate and write bindings
     let bindings = Builder::default()
