+++ program/rust/src/lib.rs
@@ -19,6 +19,7 @@ use crate::error::{
     OracleError,
     OracleResult,
 };
+
 use crate::log::{
     post_log,
     pre_log,
@@ -91,6 +92,7 @@ pub extern "C" fn entrypoint(input: *mut u8) -> u64 {
     if let Err(error) = post_log(c_ret_val, &accounts) {
         return error.into();
     }
+
 
     if c_ret_val == SUCCESSFULLY_UPDATED_AGGREGATE {
         //0 is the SUCCESS value for solana
