+++ program/rust/src/test_oracle.rs
@@ -8,10 +8,10 @@ mod test {
         PC_MAGIC,
         PC_VERSION,
     };
+    use crate::deserialize::load_account_as;
     use crate::rust_oracle::{
         clear_account,
         init_mapping,
-        load_account_as,
     };
     use bytemuck::bytes_of;
     use solana_program::account_info::AccountInfo;
